переносит код, добавляет скрипты
