+++ index.html
@@ -226,7 +226,7 @@
       </div>
     </section>
   </template>
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
   <script src="js/functions.js"></script>
 </body>
 </html>
